requirements.txt
flask==3.0.3
google-generativeai==0.7.2